Feat: 공간 관리 API 관리자 전용 접근 및 태그 등록 API 경로 변경
- 공간 관리 API에 관리자 인증 적용
- 태그 등록 API를 공통 공간 URL에서 공간 관리 URL로 변경
